Completed build of library manifest
