+++ Makefile
@@ -75,7 +75,7 @@ copy-md-files:
 build: workspace ide_build ide_native_build
 	@echo "------- Done 'build:' -------------"
 
-workspace:
+workspace: flowstandardlib
 	@echo ""
 	@echo "------- Starting build of 'flow' workspace project -------------"
 	@cargo build --all
@@ -106,10 +106,10 @@ test-ide:
 	@cd ide && make test
 
 #################### LIBRARIES ####################
-flowstdlib:
+flowstandardlib:
 	@echo ""
 	@echo "------- Starting build of 'flowstdlib' -------------"
-	@flowc flowstdlib
+	@cargo run -p flowc -- --lib flowstdlib
 	@echo "------- Done     build of 'flowstdlib' -------------"
 
 #################### Raspberry Pi ####################
@@ -129,7 +129,7 @@ copy:
 # make paths - to compile all samples found in there. Avoid files using the filter.
 sample_flows := $(patsubst samples/%,samples/%test.output,$(filter %/, $(wildcard samples/*/)))
 
-samples: workspace flowstdlib clean-samples $(sample_flows)  # This target must be below sample-flows in the Makefile
+samples: workspace clean-samples $(sample_flows)  # This target must be below sample-flows in the Makefile
 	@echo ""
 	@echo "All local samples executed and output as expected"
 	@echo "------- Finished 'samples:' ----"
