+++ flowc/src/lib_build.rs
@@ -1,14 +1,14 @@
+use std::fs::File;
+use std::io::prelude::*;
 use std::path::PathBuf;
 
 use url::Url;
 
 use flowclib::compiler::loader;
 use flowclib::deserializers::deserializer_helper::get_deserializer;
-use flowclib::model::library::Library;
 use flowclib::model::name::HasName;
-use flowclib::model::name::Name;
 use flowclib::model::process::Process::FunctionProcess;
-use flowrlib::lib_manifest::ImplementationLocator::Wasm;
+use flowrlib::lib_manifest::DEFAULT_LIB_MANIFEST_FILENAME;
 use flowrlib::lib_manifest::LibraryManifest;
 use flowrlib::manifest::MetaData;
 use flowrlib::provider::Provider;
@@ -19,25 +19,52 @@ use crate::errors::*;
 /*
     Compile a Library
 */
-pub fn build_lib(url: Url, _provided_implementation: bool, out_dir: PathBuf, provider: &dyn Provider) -> Result<String> {
-    let library = loader::load_library(&url.to_string(), provider).expect("Could not load Library");
-    build_manifest(&library, &out_dir.to_str().unwrap(), provider).expect("Could not build library manifest");
+pub fn build_lib(url: Url, _provided_implementation: bool, lib_dir: PathBuf, provider: &dyn Provider) -> Result<String> {
+//    let lib_dir = base_dir.canonicalize().expect("Could not canonicalize the specified library path");
+    let library = loader::load_library(&url.to_string(), provider)
+        .expect(&format!("Could not load Library from '{}'", lib_dir.display()));
+
+    info!("Building manifest for '{}' in output directory: '{}'\n", library.name, lib_dir.display());
+    let mut lib_manifest = LibraryManifest::new(MetaData::from(&library));
+
+    build_manifest(&mut lib_manifest, &lib_dir.to_str().unwrap(), provider).expect("Could not build library manifest");
+
+    let filename = write_lib_manifest(&lib_manifest, lib_dir)?;
+    info!("Generated library manifest at '{}'", filename.display());
+
     Ok("ok".into())
 }
 
-fn build_manifest(library: &Library, out_dir: &str, provider: &dyn Provider) -> Result<LibraryManifest> {
-    info!("Building manifest for '{}' in output directory: '{}'\n", library.name, out_dir);
-    let mut lib_manifest = LibraryManifest::new(MetaData::from(library));
+/*
+    Generate a manifest for the library in JSON that can be used to load it using 'flowr'
+*/
+fn write_lib_manifest(lib_manifest: &LibraryManifest, base_dir: PathBuf) -> Result<PathBuf> {
+    let mut filename = base_dir.clone();
+    filename.push(DEFAULT_LIB_MANIFEST_FILENAME.to_string());
+    let mut manifest_file = File::create(&filename).chain_err(|| "Could not create lib manifest file")?;
 
-    // TODO generalize to all valid extensions
-    let search_pattern = "**/*.toml";
+    manifest_file.write_all(serde_json::to_string_pretty(lib_manifest)
+        .chain_err(|| "Could not pretty format the library manifest JSON contents")?
+        .as_bytes()).chain_err(|| "Could not write library smanifest data bytes to created manifest file")?;
+
+    Ok(filename)
+}
+
+fn build_manifest(lib_manifest: &mut LibraryManifest, base_dir: &str, provider: &dyn Provider) -> Result<()> {
+    let search_pattern = if base_dir.ends_with("/") {
+        format!("{}**/*.toml", base_dir)
+    } else {
+        format!("{}/**/*.toml", base_dir)
+    };
+
     debug!("Searching for process definitions using search pattern: '{}':\n", search_pattern);
-    for entry in glob(search_pattern).expect("Failed to read glob pattern") {
+    for entry in glob(&search_pattern).expect("Failed to read glob pattern") {
         match entry {
-            Ok(ref path) => {
-                let resolved_url = Url::from_file_path(&path)
+            Ok(ref toml_path) => {
+                let resolved_url = Url::from_file_path(&toml_path)
                     .map_err(|_| format!("Could not create url from file path '{}'",
-                                         path.to_str().unwrap()))?.to_string();
+                                         toml_path.to_str().unwrap()))?.to_string();
+                debug!("Inspecting '{}' for function definition", resolved_url);
                 let contents = provider.get(&resolved_url)
                     .chain_err(|| format!("Could not get contents of resolved url: '{}'", resolved_url))?;
                 let deserializer = get_deserializer(&resolved_url)?;
@@ -45,29 +72,23 @@ fn build_manifest(library: &Library, out_dir: &str, provider: &dyn Provider) -> 
                     Ok(process) => {
                         match process {
                             FunctionProcess(function) => {
-                                add_to_manifest(path, function.name(), &mut lib_manifest);
+                                let mut wasm_abs_path: PathBuf = toml_path.clone();
+                                wasm_abs_path.set_extension("wasm");
+                                let wasm_dir = wasm_abs_path.parent().expect("Could not get parent directory of wasm path");
+                                lib_manifest.add_to_manifest(base_dir,
+                                                             wasm_abs_path.to_str().expect("Could not convert wasm_path to str"),
+                                                             wasm_dir.to_str().expect("Could not convert wasm_dir to str"),
+                                                             function.name() as &str);
                             }
                             _ => { /* Ignore valid flow definitions */ }
                         }
                     }
-                    Err(e) => Err(e).chain_err(|| format!("Could not deserialize from file {}", resolved_url))?
+                    Err(_) => { warn!("Skipping file '{}' that could not be deserialized as a process", resolved_url) }
                 }
             }
-            Err(_) => {/* Skipping unreadable files */}
+            Err(_) => { /* Skipping unreadable files */ }
         }
     }
 
-    Ok(lib_manifest)
-}
-
-fn add_to_manifest(path: &PathBuf, function_name: &Name, manifest: &mut LibraryManifest) {
-    let subpath_str = path
-        .to_str()
-        .expect("Could not convert to str")
-        .replace(".toml", "");
-
-    let impl_reference = format!("//{}/{}/{}", manifest.metadata.name, subpath_str, function_name);
-    debug!("Adding function to manifest: '{}'", impl_reference);
-    let implementation_location = format!("");
-    manifest.locators.insert(impl_reference, Wasm(implementation_location));
+    Ok(())
 }
