+++ flowc/tests/test_libs/Library_test.toml
@@ -1,4 +1,4 @@
-library = "flowstdlib"
+library = "Flow Standard Library"
 version = "0.7.0"
 author_name = "Andrew Mackenzie"
 author_email = "<andrew@mackenzie-serres.net>"
