+++ flowstdlib/.gitignore
@@ -1,1 +1,2 @@
 Cargo.lock
+manifest.json
