--- flowstdlib/manifest.json
@@ -1,24 +1,0 @@
-{
-  "metadata": {
-    "name": "Flow Standard Library",
-    "version": "0.7.0",
-    "author_name": "Andrew Mackenzie",
-    "author_email": "<andrew@mackenzie-serres.net>",
-    "description": "The standard library for 'flow' programs compiled with the 'flowc' compiler"
-  },
-  "locators": {
-    "lib://flowstdlib/control/compare/Compare": "control/compare/compare.wasm",
-    "lib://flowstdlib/control/join/Join": "control/join/join.wasm",
-    "lib://flowstdlib/control/tap/Tap": "control/tap/tap.wasm",
-    "lib://flowstdlib/data/buffer/Buffer": "data/buffer/buffer.wasm",
-    "lib://flowstdlib/data/compose_array/ComposeArray": "data/compose_array/compose_array.wasm",
-    "lib://flowstdlib/data/zip/Zip": "data/zip/zip.wasm",
-    "lib://flowstdlib/fmt/reverse/Reverse": "fmt/reverse/reverse.wasm",
-    "lib://flowstdlib/fmt/to_string/ToString": "fmt/to_string/to_string.wasm",
-    "lib://flowstdlib/fmt/to_number/ToNumber": "fmt/to_number/to_number.wasm",
-    "lib://flowstdlib/img/format_png/FormatPNG": "img/format_png/format_png.wasm",
-    "lib://flowstdlib/math/add/Add": "math/add/add.wasm",
-    "lib://flowstdlib/math/divide/Divide": "math/divide/divide.wasm",
-    "lib://flowstdlib/math/subtract/Subtract": "math/subtract/subtract.wasm"
-  }
-}
